finish making hero, role and tech reusable
